+++ JumpScale9/clients/tarantool/TarantoolDB.py
@@ -10,8 +10,11 @@ class TarantoolDB():
     def __init__(self, name="test", path="$DATADIR/tarantool/$NAME", adminsecret="admin007", port=3301):
         self.path = j.dirs.replaceTxtDirVars(
             path).replace("$NAME", name).strip()
+        j.sal.fs.createDir(self.path)
         self.name = name
+        self.login="root"
         self.adminsecret = adminsecret
+        self.addr="localhost"
         self.port = port
         self.configTemplatePath = None
 
@@ -38,13 +41,13 @@ class TarantoolDB():
         j.sal.fs.writeFile(j.clients.tarantool.cfgdir +
                            "/%s.lua" % self.name, C2)
 
-    def startConnect(self):
+    def start_connect(self):
         """
         will start a local tarantool in console
         """
         self._setConfig()
-        j.do.execute("tarantoolctl start test")
-        j.do.executeInteractive("tarantoolctl enter test")
+        j.do.execute("tarantoolctl start %s"%self.name)
+        j.do.executeInteractive("tarantoolctl enter %s"%self.name)
 
         # FOR TEST PURPOSES (DEBUG ON CONSOLE)
         # rm 000*;rm -rf /Users/kristofdespiegeleer1/opt/var/data/tarantool/test;tarantoolctl start test; cat /Users/kristofdespiegeleer1/opt/var/data/tarantool/test/instance.log
@@ -52,9 +55,14 @@ class TarantoolDB():
     def start(self):
         # j.tools.prefab.local.db.tarantool.start()
         self._setConfig()
-        j.do.execute("tarantoolctl start test")
-        j.do.executeInteractive("tarantoolctl enter test")
+        j.do.execute("tarantoolctl start %s"%self.name)
+        # j.do.executeInteractive("tarantoolctl enter %s"%self.name)
 
-    def connect(self, addr="localhost", port=3301):
-        cmd = "tarantoolctl connect root:admin007@%s:%s" % (addr, port)
+    def connect_shell(self):
+        """
+        connect over tcp to the running tarantool
+        """
+        cmd = "tarantoolctl connect %s:%s@%s:%s" % (self.login,self.adminsecret,self.addr, self.port)
         j.do.executeInteractive(cmd)
+
+
