+++ JumpScale9/data/key_value_store/memory_store.py
@@ -84,7 +84,6 @@ class MemoryKeyValueStore(KeyValueStoreBase):
             indexitem is e.g. $actorname:$state:$role (is a text which will be index to key)
                 indexitems are always made lowercase
             key links to the object in the db
-        ':' is not allowed in indexitem
         """
         self.dbindex.update(items)
